refactor: mv type
