+++ src/components/Compose/ReplyRestriction.tsx
@@ -5,10 +5,9 @@ import { Fragment } from 'react';
 import RadioDisableNoIcon from '@/assets/radio.disable-no.svg';
 import YesIcon from '@/assets/yes.svg';
 import { ReplyRestrictionText } from '@/components/Compose/ReplyRestrictionText.js';
-import { SocialPlatform } from '@/constants/enum.js';
+import { RestrictionType, SocialPlatform } from '@/constants/enum.js';
 import { classNames } from '@/helpers/classNames.js';
 import type { CompositePost } from '@/store/useComposeStore.js';
-import { RestrictionType } from '@/types/compose.js';
 
 interface ReplyRestrictionProps {
     post: CompositePost;
