+++ src/constants/index.ts
@@ -1,8 +1,7 @@
 /* cspell:disable */
 
-import { NODE_ENV, ProfileTabType, SearchType, SocialPlatform, VERCEL_NEV } from '@/constants/enum.js';
+import { NODE_ENV, ProfileTabType, RestrictionType, SearchType, SocialPlatform, VERCEL_NEV } from '@/constants/enum.js';
 import { env } from '@/constants/env.js';
-import { RestrictionType } from '@/types/compose.js';
 
 export const SITE_NAME = 'Firefly: Web3 & NFT Explorer';
 export const SITE_DESCRIPTION =
@@ -22,12 +21,6 @@ export const RP_HASH_TAG = '#FireflyLuckyDrop';
 
 export const EMPTY_LIST = Object.freeze([]) as never[];
 export const EMPTY_OBJECT = Object.freeze({}) as Record<string, never>;
-
-export const EIP_712_FARCASTER_DOMAIN = {
-    name: 'Farcaster Verify Ethereum Address',
-    version: '2.0.0',
-    salt: '0xf2d857f4a3edcb9b78b4d503bfe733db1e3f6cdc2b7971ee739626c97e86a558' as `0x${string}`,
-};
 
 export const SORTED_PROFILE_TAB_TYPE = [ProfileTabType.Feed, ProfileTabType.Collected, ProfileTabType.Channels];
 export const SORTED_SEARCH_TYPE = [SearchType.Posts, SearchType.Users];
