+++ src/types/compose.ts
@@ -1,7 +1,1 @@
 export type ComposeType = 'compose' | 'quote' | 'reply';
-
-export enum RestrictionType {
-    Everyone = 0,
-    OnlyPeopleYouFollow = 1,
-    MentionedUsers = 2,
-}
